requirements.txt
py-cord==2.2.2
python-dotenv==0.21.0
requests==2.28.1
